feat(tree): enhance tree component functionality and structure
