+++ src/renderer/components/ui/tree/index.ts
@@ -1,1 +1,3 @@
 export { default as Tree } from './Tree.vue'
+
+export * from './types'
